refactoring tests with mock servers (#30) (#356)
